chore(quality): resolve slop-detector issues #504–#508
Strengthen prod-only slop scans, split high-complexity agent/recall paths,
and document TS gate HALT limits while keeping CI advisory.
