+++ .github/workflows/slop-detector-js.yml
@@ -1,5 +1,6 @@
-# Advisory JS/TS scan via ai-slop-detector. Not a required merge check (#314).
-# Triggers: manual + weekly schedule. Logs: JS/TS Clean/Suspicious/Critical (ignore --gate HALT for TS-only context).
+# Advisory JS/TS scan via ai-slop-detector. Not a required merge check (#314, #504).
+# Triggers: manual + weekly schedule. Project gate HALT (LDR/DDC) is expected for TS monorepos;
+# use per-file Clean/Suspicious/Critical counts and targeted refactors as the actionable signal.
 
 name: Slop detector (JS/TS)
 
@@ -30,12 +31,15 @@ jobs:
           python-version: '3.12'
 
       - name: Install ai-slop-detector
-        run: pip install --progress-bar off 'ai-slop-detector==3.7.3'
+        run: pip install --progress-bar off 'ai-slop-detector==3.8.5'
 
-      - name: Scan packages/ (JS/TS)
+      - name: Scan packages (prod src, JS/TS gate report)
         run: |
           set -o pipefail
-          slop-detector --project packages --js --config .slopconfig.yaml --no-color 2>&1 | tee slop-packages.log
+          for pkg in memento-core memento-server memento-client; do
+            echo "=== packages/$pkg/src ==="
+            slop-detector --project "packages/$pkg/src" --js --config .slopconfig.yaml --gate --no-color 2>&1 | tee -a slop-packages-gate.log
+          done
 
       - name: Scan static/js (JS/TS)
         run: |
