+++ .slopconfig.yaml
@@ -1,5 +1,13 @@
 version: "2.0"
 ignore:
+  - "**/*.spec.ts"
+  - "**/*.spec.js"
+  - "**/*.test.ts"
+  - "**/__tests__/**"
+  - "**/test/**"
+  - "tests/**"
+  - "packages/**/src/test/**"
+  - "packages/**/src/**/__tests__/**"
   - "**/coverage/**"
   - "graphify-out/**"
   - "**/.next/**"
