+++ packages/memento-core/src/domains/memory/tools/recall-tool-envelope.ts
@@ -7,13 +7,12 @@ import { INTROSPECTION_HINT_SUFFIX } from '../../../shared/constants/introspecti
 import type { EmbeddingProvider, MemorySearchFilters } from '../../../shared/types/index.js';
 import { emitTfidfFallbackWarningIfNeeded } from '../../../shared/utils/embedding-provider-diagnostics.js';
 import type { ToolContext, ToolResult } from '../../../tools/types.js';
-import { getVectorSearchEngine } from '../../search/algorithms/vector-search-engine.js';
 import type { NeighborMemory } from '../services/memory-neighbor-service.js';
-import { MemoryEmbeddingService } from '../services/memory-embedding-service.js';
-import { MemoryNeighborService } from '../services/memory-neighbor-service.js';
 import type { MetaMemoryService } from '../services/meta-memory-service.js';
 import { getAppliedRecallFilters } from './recall-tool-filters.js';
+import { handleAutoSetAnchor } from './recall-tool-anchor-rotation.js';
 import type { RecallToolHost } from './recall-tool-host.js';
+import { handleIncludeNeighbors } from './recall-tool-neighbors-fetch.js';
 import {
   buildQueryEmbeddingMetadataFields,
   recallQueryCorrelationExtra
@@ -72,266 +71,6 @@ export async function getMetaStatsForResults(
       items_count: processedResults.length
     });
     return undefined;
-  }
-}
-
-/**
- * 자동 앵커 설정 처리
- */
-export async function handleAutoSetAnchor(
-  host: RecallToolHost,
-  searchItems: RecallSearchItem[],
-  agentId: string,
-  context: ToolContext
-): Promise<{
-  success: boolean;
-  anchor_set: AnchorSetMetadata | null;
-  error?: boolean;
-  skipped?: boolean;
-  skipped_reason?: string;
-}> {
-  if (!searchItems || searchItems.length === 0) {
-    return {
-      success: false,
-      anchor_set: null
-    };
-  }
-
-  const topMemory = searchItems[0]!;
-  const memoryId = topMemory.id ?? topMemory.memory_id;
-
-  if (!memoryId) {
-    host.logWarning('검색 결과에 memory_id가 없어 앵커 설정을 건너뜁니다', { topMemory });
-    return {
-      success: false,
-      anchor_set: null,
-      error: true
-    };
-  }
-
-  if (!context.services.anchorManager) {
-    host.logWarning('AnchorManager 서비스가 없어 앵커 설정을 건너뜁니다');
-    return {
-      success: false,
-      anchor_set: null,
-      error: true
-    };
-  }
-
-  try {
-    const slotAAnchor = await context.services.anchorManager.getAnchor(agentId, 'A');
-
-    if (slotAAnchor && typeof slotAAnchor === 'object' && 'memory_id' in slotAAnchor) {
-      const anchorMemory = context.db!.prepare(`
-          SELECT pinned FROM memory_item WHERE id = ?
-        `).get(slotAAnchor.memory_id) as { pinned: number | boolean } | undefined;
-
-      const isPinned = anchorMemory && (anchorMemory.pinned === 1 || anchorMemory.pinned === true);
-
-      if (isPinned) {
-        host.logInfo('슬롯 A에 pinned 앵커가 있어 앵커 설정을 건너뜁니다', {
-          agent_id: agentId,
-          existing_memory_id: slotAAnchor.memory_id
-        });
-        return {
-          success: false,
-          anchor_set: null,
-          skipped: true,
-          skipped_reason: 'pinned_anchor_protected'
-        };
-      }
-
-      const slotBAnchor = await context.services.anchorManager.getAnchor(agentId, 'B');
-
-      if (slotBAnchor && typeof slotBAnchor === 'object' && 'memory_id' in slotBAnchor) {
-        const slotBMemory = context.db!.prepare(`
-            SELECT pinned FROM memory_item WHERE id = ?
-          `).get(slotBAnchor.memory_id) as { pinned: number | boolean } | undefined;
-
-        const slotBIsPinned = slotBMemory && (slotBMemory.pinned === 1 || slotBMemory.pinned === true);
-
-        if (slotBIsPinned) {
-          host.logWarning('슬롯 B의 pinned 앵커가 덮어써집니다', {
-            agent_id: agentId,
-            old_memory_id: slotBAnchor.memory_id,
-            new_memory_id: slotAAnchor.memory_id
-          });
-        }
-
-        const slotCAnchor = await context.services.anchorManager.getAnchor(agentId, 'C');
-
-        if (slotCAnchor && typeof slotCAnchor === 'object' && 'memory_id' in slotCAnchor) {
-          const slotCMemory = context.db!.prepare(`
-              SELECT pinned FROM memory_item WHERE id = ?
-            `).get(slotCAnchor.memory_id) as { pinned: number | boolean } | undefined;
-
-          const slotCIsPinned = slotCMemory && (slotCMemory.pinned === 1 || slotCMemory.pinned === true);
-
-          if (slotCIsPinned) {
-            host.logWarning('슬롯 C의 pinned 앵커가 제거됩니다', {
-              agent_id: agentId,
-              old_memory_id: slotCAnchor.memory_id
-            });
-          }
-
-          await context.services.anchorManager.clearAnchor(agentId, 'C');
-        }
-
-        const slotBMemoryId = slotBAnchor.memory_id;
-        if (slotBMemoryId) {
-          await context.services.anchorManager.clearAnchor(agentId, 'B');
-          await context.services.anchorManager.setAnchor(agentId, slotBMemoryId, 'C');
-        }
-      }
-
-      const slotAMemoryId = slotAAnchor.memory_id;
-      if (slotAMemoryId) {
-        await context.services.anchorManager.clearAnchor(agentId, 'A');
-        await context.services.anchorManager.setAnchor(agentId, slotAMemoryId, 'B');
-      }
-    }
-
-    await context.services.anchorManager.setAnchor(agentId, memoryId, 'A');
-
-    host.logInfo('앵커가 자동으로 설정되었습니다', {
-      agent_id: agentId,
-      memory_id: memoryId,
-      slot: 'A'
-    });
-
-    return {
-      success: true,
-      anchor_set: {
-        memory_id: memoryId,
-        slot: 'A',
-        agent_id: agentId
-      }
-    };
-  } catch (error) {
-    host.logError(error as Error, '앵커 자동 설정 실패', {
-      agent_id: agentId,
-      memory_id: memoryId
-    });
-
-    return {
-      success: false,
-      anchor_set: null,
-      error: true
-    };
-  }
-}
-
-/**
- * 자동 이웃 기억 포함 처리
- */
-export async function handleIncludeNeighbors(
-  host: RecallToolHost,
-  searchItems: RecallSearchItem[],
-  neighborsLimit: number,
-  neighborsPerItem: number,
-  neighborsSimilarityThreshold: number,
-  context: ToolContext
-): Promise<NeighborMemory[][]> {
-  if (!searchItems || searchItems.length === 0) {
-    return [];
-  }
-
-  const topResults = searchItems.slice(0, Math.min(neighborsLimit, searchItems.length));
-
-  let neighborService: MemoryNeighborService;
-  try {
-    const vectorSearchEngine = context.services?.vectorSearchEngine ?? getVectorSearchEngine();
-    const embeddingService = context.services.embeddingService || new MemoryEmbeddingService();
-    neighborService = new MemoryNeighborService(vectorSearchEngine, embeddingService, context.db!);
-  } catch (error) {
-    host.logError(error as Error, 'MemoryNeighborService 초기화 실패', {});
-    return Array.from({ length: topResults.length }, () => []);
-  }
-
-  const neighborPromises = topResults.map(async (item, index) => {
-    const memoryId = item.id || item.memory_id;
-
-    if (!memoryId) {
-      host.logWarning('검색 결과에 memory_id가 없어 이웃 기억 조회를 건너뜁니다', { item });
-      return { index, neighbors: [] };
-    }
-
-    try {
-      const timeoutPromise = new Promise<{ index: number; neighbors: NeighborMemory[] }>((_, reject) => {
-        setTimeout(() => reject(new Error('Timeout')), 2000);
-      });
-
-      const neighborPromise = neighborService.getNeighbors(memoryId, {
-        limit: neighborsPerItem,
-        similarity_threshold: neighborsSimilarityThreshold
-      }).then(result => ({
-        index,
-        neighbors: result.neighbors
-      }));
-
-      const result = await Promise.race([neighborPromise, timeoutPromise]);
-      return result;
-    } catch (error) {
-      if (error instanceof Error && error.message === 'Timeout') {
-        host.logWarning('이웃 기억 조회 타임아웃', { memoryId, index });
-      } else {
-        host.logError(error as Error, '이웃 기억 조회 실패', { memoryId, index });
-      }
-      return { index, neighbors: [] };
-    }
-  });
-
-  const completedResults = new Map<number, { index: number; neighbors: NeighborMemory[] }>();
-
-  neighborPromises.forEach((promise, idx) => {
-    promise
-      .then(result => {
-        completedResults.set(idx, result);
-      })
-      .catch(() => {
-        completedResults.set(idx, { index: idx, neighbors: [] });
-      });
-  });
-
-  let timeoutId: NodeJS.Timeout | null = null;
-  const timeoutPromise = new Promise<Array<{ index: number; neighbors: NeighborMemory[] }>>((resolve) => {
-    timeoutId = setTimeout(() => {
-      const partialResults: Array<{ index: number; neighbors: NeighborMemory[] }> = [];
-
-      for (let i = 0; i < topResults.length; i++) {
-        if (completedResults.has(i)) {
-          partialResults.push(completedResults.get(i)!);
-        } else {
-          partialResults.push({ index: i, neighbors: [] });
-        }
-      }
-
-      resolve(partialResults.sort((a, b) => a.index - b.index));
-    }, 2500);
-  });
-
-  try {
-    const allNeighbors = await Promise.race([
-      Promise.all(neighborPromises),
-      timeoutPromise
-    ]);
-
-    if (timeoutId) clearTimeout(timeoutId);
-
-    const sortedNeighbors = allNeighbors
-      .sort((a, b) => a.index - b.index)
-      .map(r => r.neighbors);
-
-    return sortedNeighbors;
-  } catch (error) {
-    if (timeoutId) clearTimeout(timeoutId);
-
-    const settledResults = await Promise.allSettled(neighborPromises);
-    return settledResults.map((r, _idx) =>
-      r.status === 'fulfilled'
-        ? r.value.neighbors
-        : []
-    );
   }
 }
